rename(#88): 서비스를 product, category 패키지로 구분
