+++ service/product/server/src/main/java/com/sparta/product/application/product/ProductMapper.java
@@ -1,4 +1,4 @@
-package com.sparta.product.application;
+package com.sparta.product.application.product;
 
 import com.sparta.product.domain.model.Product;
 import com.sparta.product.presentation.request.ProductCreateRequest;
@@ -54,5 +54,4 @@ public class ProductMapper {
         request.tags(),
         request.isPublic());
   }
-
 }
